requirements.txt
Flask>=3.0
opencv-python-headless>=4.8
Pillow>=10.0
numpy>=1.24
reportlab>=4.0
gunicorn>=21.2
